+++ CHANGELOG.md
@@ -26,6 +26,7 @@
   - Fix to preserve kebab-casing in server-info-raw. #1195
   - Add refactoring `Restructure keys`, the inverse of `Destructure keys`. #1170
   - Add refactorings to convert between `(:x m)` and `(get m :x)` or `(:y (:x m))` and `(get-in m [:x :y])`. #1172
+  - Add support to imported java class on completion. #1193
 
 ## 2022.07.24-18.25.43